Add FileParserBase class
